Fix test code to better reuse function trigger events from evetns folder
